+++ packages/lisk-passphrase/src/index.ts
